requirements.txt
rpi.gpio
pandas
glob2
tables
pdfminer.six
